cayvibes-api-gateway/requirements.txt
Django~=5.0.6
djangorestframework==3.15.1
django-cors-headers==3.14.0
requests==2.32.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg[binary]==3.1.19
drf-social-oauth2==2.2.0
<<<<<<< HEAD
pika==1.3.2
django-countries==7.6.1
django-phonenumber-field[phonenumbers]
=======
django-countries==7.6.1
pika==1.3.2
>>>>>>> d0568b6 (adding files)
